+++ packages/atlas/src/App.tsx
@@ -1,7 +1,7 @@
 import { AnalyticsManager } from '@/AnalyticsManager'
 import { CommonProviders } from '@/CommonProviders'
 import { WelcomeDialog } from '@/components/WelcomeDialog'
-import { SignInModal } from '@/components/_auth/SignInModal'
+import { ExternalSignInModal } from '@/components/_auth/ExternalSignInModal'
 import { JoystreamManager } from '@/providers/joystream/joystream.manager'
 import { JoystreamProvider } from '@/providers/joystream/joystream.provider'
 import { NftActionsProvider } from '@/providers/nftActions/nftActions.provider'
@@ -25,7 +25,7 @@ export const App = () => {
           <TransactionsManager />
           <JoystreamManager />
           <NotificationsManager />
-          <SignInModal />
+          <ExternalSignInModal />
           <WelcomeDialog />
           <NftSettlementBottomDrawer />
           <NftPurchaseBottomDrawer />
