+++ packages/atlas/src/components/_auth/ExternalSignInModal/ExternalSignInModal.styles.ts
