+++ packages/atlas/src/components/_auth/ExternalSignInModal/ExternalSignInSteps/ExternalSignInModalStepTemplate/ExternalSignInModalStepTemplate.styles.ts
