+++ packages/atlas/src/components/_auth/ExternalSignInModal/ExternalSignInSteps/ExternalSignInModalStepTemplate/ExternalSignInModalStepTemplate.tsx
@@ -13,7 +13,7 @@ import {
   HeaderContainer,
   LogoContainer,
   StyledAppLogo,
-} from './SignInModalStepTemplate.styles'
+} from './ExternalSignInModalStepTemplate.styles'
 
 type SignInModalStepTemplateProps = PropsWithChildren<{
   title: string
@@ -25,7 +25,7 @@ type SignInModalStepTemplateProps = PropsWithChildren<{
   formNode?: ReactNode
 }>
 
-export const SignInModalStepTemplate: FC<SignInModalStepTemplateProps> = ({
+export const ExternalSignInModalStepTemplate: FC<SignInModalStepTemplateProps> = ({
   title,
   subtitle,
   children,
