+++ packages/atlas/src/components/_auth/ExternalSignInModal/ExternalSignInSteps/ExternalSignInModalWalletStep/index.ts
@@ -1,0 +1,1 @@
+export * from './ExternalSignInModalWalletStep'
