+++ packages/atlas/src/components/_auth/ExternalSignInModal/ExternalSignInSteps/ExternalSignInSteps.styles.ts
