+++ packages/atlas/src/components/_auth/ExternalSignInModal/ExternalSignInSteps/index.ts
@@ -1,0 +1,4 @@
+export * from './ExternalSignInModalWalletStep/ExternalSignInModalWalletStep'
+export * from './ExternalSignInModalMembershipsStep'
+export type { SignInStepProps } from './ExternalSignInSteps.types'
+export { ModalSteps } from './ExternalSignInSteps.types'
