+++ packages/atlas/src/components/_auth/ExternalSignInModal/index.ts
@@ -1,0 +1,1 @@
+export * from './ExternalSignInModal'
