--- packages/atlas/src/components/_auth/SignInModal/SignInSteps/SignInModalStepTemplate/index.ts
@@ -1,1 +1,0 @@
-export * from './SignInModalStepTemplate'
