--- packages/atlas/src/components/_auth/SignInModal/SignInSteps/SignInModalWalletStep/index.ts
@@ -1,1 +1,0 @@
-export * from './SignInModalWalletStep'
