--- packages/atlas/src/components/_auth/SignInModal/SignInSteps/index.ts
@@ -1,4 +1,0 @@
-export * from './SignInModalWalletStep/SignInModalWalletStep'
-export * from './SignInModalMembershipsStep'
-export type { SignInStepProps } from './SignInSteps.types'
-export { ModalSteps } from './SignInSteps.types'
