--- packages/atlas/src/components/_auth/SignInModal/index.ts
@@ -1,1 +1,0 @@
-export * from './SignInModal'
